+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - DocumentDb: Added operator Neq (Fix #48)
 - DocumentDb: Added option overwrite to set to allow pure create operation (Fix #52)
 - Templates: Support translations/resources (Fix #32)
+- Custom payload for errors + standard oauthError (Fix #56)
 - Body validation and class assignment in HttpUnoEvent
 
 ### Changed