+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.48.0] - 2018-08-23
 ### Added
 - DocumentDb: Added operator Neq (Fix #48)
 - DocumentDb: Added option overwrite to set to allow pure create operation (Fix #52)